+++ src/api_server/src/lib.rs
@@ -107,6 +107,7 @@ impl ApiServer {
     /// # Example
     ///
     /// ```
+    /// use mmds::MAX_DATA_STORE_SIZE;
     /// use api_server::ApiServer;
     /// use mmds::MMDS;
     /// use std::{
@@ -130,6 +131,7 @@ impl ApiServer {
     /// let mmds_info = MMDS.clone();
     /// let time_reporter = ProcessTimeReporter::new(Some(1), Some(1), Some(1));
     /// let seccomp_filters = get_filters(SeccompConfig::None).unwrap();
+    /// let payload_limit = Some(MAX_DATA_STORE_SIZE);
     ///
     /// thread::Builder::new()
     ///     .name("fc_api_test".to_owned())
@@ -144,6 +146,7 @@ impl ApiServer {
     ///             PathBuf::from(api_thread_path_to_socket),
     ///             time_reporter,
     ///             seccomp_filters.get("api").unwrap(),
+    ///             payload_limit,
     ///         )
     ///         .unwrap();
     ///     })
@@ -166,11 +169,15 @@ impl ApiServer {
         path: PathBuf,
         process_time_reporter: ProcessTimeReporter,
         seccomp_filter: BpfProgramRef,
+        maybe_request_size: Option<usize>,
     ) -> Result<()> {
         let mut server = HttpServer::new(path).unwrap_or_else(|e| {
             error!("Error creating the HTTP server: {}", e);
             std::process::exit(vmm::FC_EXIT_CODE_GENERIC_ERROR);
         });
+        if let Some(request_size) = maybe_request_size {
+            server.set_payload_max_size(request_size);
+        }
 
         // Store process start time metric.
         process_time_reporter.report_start_time();
@@ -320,6 +327,10 @@ impl ApiServer {
                 data_store::Error::UnsupportedValueType => unreachable!(),
                 data_store::Error::NotInitialized => ApiServer::json_response(
                     StatusCode::BadRequest,
+                    ApiServer::json_fault_message(e.to_string()),
+                ),
+                data_store::Error::DataStoreLimitExceeded => ApiServer::json_response(
+                    StatusCode::PayloadTooLarge,
                     ApiServer::json_fault_message(e.to_string()),
                 ),
             },
@@ -641,6 +652,7 @@ mod tests {
                     PathBuf::from(api_thread_path_to_socket),
                     ProcessTimeReporter::new(Some(1), Some(1), Some(1)),
                     seccomp_filters.get("api").unwrap(),
+                    None,
                 )
                 .unwrap();
             })
@@ -659,5 +671,61 @@ mod tests {
         assert!(sock.write_all(b"OPTIONS / HTTP/1.1\r\n\r\n").is_ok());
         let mut buf: [u8; 100] = [0; 100];
         assert!(sock.read(&mut buf[..]).unwrap() > 0);
+    }
+
+    #[test]
+    fn test_bind_and_run_with_limit() {
+        let mut tmp_socket = TempFile::new().unwrap();
+        tmp_socket.remove().unwrap();
+        let path_to_socket = tmp_socket.as_path().to_str().unwrap().to_owned();
+        let api_thread_path_to_socket = path_to_socket.clone();
+
+        let to_vmm_fd = EventFd::new(libc::EFD_NONBLOCK).unwrap();
+        let (api_request_sender, _from_api) = channel();
+        let (_to_api, vmm_response_receiver) = channel();
+        let mmds_info = MMDS.clone();
+        let seccomp_filters = get_filters(SeccompConfig::Advanced).unwrap();
+
+        thread::Builder::new()
+            .name("fc_api_test".to_owned())
+            .spawn(move || {
+                ApiServer::new(
+                    mmds_info,
+                    api_request_sender,
+                    vmm_response_receiver,
+                    to_vmm_fd,
+                )
+                .bind_and_run(
+                    PathBuf::from(api_thread_path_to_socket),
+                    ProcessTimeReporter::new(Some(1), Some(1), Some(1)),
+                    seccomp_filters.get("api").unwrap(),
+                    Some(50),
+                )
+                .unwrap();
+            })
+            .unwrap();
+
+        // Wait for the server to set itself up.
+        thread::sleep(Duration::from_millis(10));
+        let mut sock = UnixStream::connect(PathBuf::from(path_to_socket)).unwrap();
+
+        // Send a GET mmds request.
+        assert!(sock
+            .write_all(
+                b"PUT http://localhost/home HTTP/1.1\r\n\
+                  Content-Length: 50000\r\n\r\naaaaaa"
+            )
+            .is_ok());
+        let mut buf: [u8; 265] = [0; 265];
+        assert!(sock.read(&mut buf[..]).unwrap() > 0);
+        let error_message = b"HTTP/1.1 400 \r\n\
+                              Server: Firecracker API\r\n\
+                              Connection: keep-alive\r\n\
+                              Content-Type: application/json\r\n\
+                              Content-Length: 146\r\n\r\n{ \"error\": \"\
+                              Request payload with size 50000 is larger than \
+                              the limit of 50 allowed by server.\nAll previous \
+                              unanswered requests will be dropped.\" }";
+        assert_eq!(&buf[..], &error_message[..]);
     }
 }