+++ src/dumbo/src/tcp/endpoint.rs
@@ -312,7 +312,7 @@ fn build_response(http_version: Version, status_code: StatusCode, body: Body) ->
 
 /// Parses the request bytes and builds a `micro_http::Response` by the given callback function.
 fn parse_request_bytes(byte_stream: &[u8], callback: fn(Request) -> Response) -> Response {
-    let request = Request::try_from(byte_stream);
+    let request = Request::try_from(byte_stream, None);
     match request {
         Ok(request) => callback(request),
         Err(e) => match e {
@@ -359,6 +359,11 @@ fn parse_request_bytes(byte_stream: &[u8], callback: fn(Request) -> Response) ->
             RequestError::Underflow => build_response(
                 Version::default(),
                 StatusCode::BadRequest,
+                Body::new(e.to_string()),
+            ),
+            RequestError::SizeLimitExceeded(_, _) => build_response(
+                Version::default(),
+                StatusCode::PayloadTooLarge,
                 Body::new(e.to_string()),
             ),
         },