+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `--http_api_max_payload_size` parameter to configure the maximum payload
   size for PUT and PATCH requests.
+- Limit MMDS data store size to `--http_api_max_payload_size`.
 
 ## [0.25.0]