+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `--http_api_max_payload_size` parameter to configure the maximum payload
   size for PUT and PATCH requests.
 - Limit MMDS data store size to `--http_api_max_payload_size`.
+- Cleanup all environment variables in Jailer.
 
 ## [0.25.0]