+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   size for PUT and PATCH requests.
 - Limit MMDS data store size to `--http_api_max_payload_size`.
 - Cleanup all environment variables in Jailer.
+- Added metrics for accesses to deprecated HTTP and command line API endpoints.
 
 ## [0.25.0]