+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### Added
 
-- Added `--http_api_max_payload_size` parameter to configure the maximum payload
+- Added `--http-api-max-payload-size` parameter to configure the maximum payload
   size for PUT and PATCH requests.
-- Limit MMDS data store size to `--http_api_max_payload_size`.
+- Limit MMDS data store size to `--http-api-max-payload-size`.
 - Cleanup all environment variables in Jailer.
 
 ## [0.25.0]